Portable Encapsulated Project object

Provides an in-memory representation and functions to access project configuration and sample annotation values for a PEP.

Details

Can be created with the constructor: "Project"

Slots

file

character vector path to config file on disk.

samples

a data table object holding the sample metadata

config

a list object holding contents of the config file

sampleNameAttr

a string indicating the sample attribute that is used to index the sample table

subSampleNameAttr

a string indicating the sample attribute that is used to index the sample table