Activate other subproject in objects of Project class

This method switches the between the subprojects within the Project object

# S4 method for Project,character
activateSubproject(.Object, sp, ...)

Arguments

.Object
an object of class Project

sp
character with the subproject name

Details

To check what are the subproject names call listSubprojects(config(p)), where p is the object of Project class

Examples